crda: Makefile: don't compile keys-ssl.c
keys-ssl.c is included in reglib.c, so no need
to compile it explicitly (this actually causes
compilation error as BN_ULONG is not defined
in keys-ssl.c itself.

Solve it by making reglib.c (instead of LIBREG)
dependent on keys-ssl.c, and while on it
change LIBREG compilation to compile only
the relevant .c file.

Signed-off-by: Eliad Peller <eliad@wizery.com>
Signed-off-by: Guy Mishol <guym@ti.com>